refactor: extract all slash commands into standalone modules
Move all 24 slash command handlers from the MastraTUI God class into
individual files under tui/commands/. Each handler receives a
SlashCommandContext for dependency injection instead of reaching into
class internals.

mastra-tui.ts: 4,617 → 3,224 lines (-30%)
commands/: 25 new files, 1,518 lines total
