+++ mastracode/src/tui/commands/exit.ts
@@ -1,0 +1,6 @@
+import type { SlashCommandContext } from './types.js';
+
+export function handleExitCommand(ctx: SlashCommandContext): void {
+  ctx.stop();
+  process.exit(0);
+}
