+++ mastracode/src/tui/commands/index.ts
@@ -1,0 +1,25 @@
+export type { SlashCommandContext } from './types.js';
+
+export { handleHelpCommand } from './help.js';
+export { handleCostCommand } from './cost.js';
+export { handleYoloCommand } from './yolo.js';
+export { handleThinkCommand } from './think.js';
+export { handlePermissionsCommand } from './permissions.js';
+export { handleNameCommand } from './name.js';
+export { handleExitCommand } from './exit.js';
+export { handleHooksCommand } from './hooks.js';
+export { handleMcpCommand } from './mcp.js';
+export { handleModeCommand } from './mode.js';
+export { handleSkillsCommand } from './skills.js';
+export { handleNewCommand } from './new.js';
+export { handleResourceCommand } from './resource.js';
+export { handleDiffCommand } from './diff.js';
+export { handleThreadsCommand } from './threads.js';
+export { handleThreadTagDirCommand } from './thread-tag-dir.js';
+export { handleSandboxCommand } from './sandbox.js';
+export { handleModelsCommand } from './models.js';
+export { handleSubagentsCommand } from './subagents.js';
+export { handleOMCommand } from './om.js';
+export { handleSettingsCommand } from './settings.js';
+export { handleLoginCommand } from './login.js';
+export { handleReviewCommand } from './review.js';
